split the GUI into a new package
gui.py was far too large and needed splitting up to make it
easier to maintain. This commit creates a new "gui" package
and puts each panel into its own module. The worker threads are
all combined into one module, which might need to change in the
future, but given their current size is probably ok.
